requirements.txt
blinker==1.9.0
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-WTF==1.2.2
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
MarkupSafe==3.0.3
narwhals==2.8.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
plotly==6.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
